feat(kap-server): add remote-control runtime toggle API
