+++ apps/kimi-code/package.json
@@ -91,6 +91,7 @@
     "@moonshot-ai/migration-legacy": "workspace:^",
     "@moonshot-ai/minidb": "workspace:^",
     "@moonshot-ai/pi-tui": "workspace:^",
+    "@moonshot-ai/remote-control": "workspace:^",
     "@moonshot-ai/vis-server": "workspace:^",
     "@moonshot-ai/vis-web": "workspace:*",
     "@types/qrcode": "^1.5.6",
