+++ apps/kimi-code/src/cli/sub/web/run.ts
@@ -224,6 +224,7 @@ export async function handleWebCommand(
           homeDir: dataDir,
           localOrigin: origin,
           localServerToken: token,
+          clientVersion: `kimi-code/${getVersion()}`,
           stderr: deps.stderr,
           onStatus,
         });
