+++ apps/kimi-code/src/tui/commands/web.ts
@@ -1,17 +1,16 @@
 import chalk from 'chalk';
 
 import { splitTokenFragment } from '#/cli/sub/web/access-urls';
+import { getVersion } from '#/cli/version';
 import {
   buildRemoteControlUrl,
+  formatRemoteControlAlreadyRunning,
   formatRemoteControlOutput,
   formatRemoteControlStatus,
+  inspectRemoteControlLock,
   startRemoteControl,
   type RemoteControlStatus,
 } from '#/cli/sub/web/remote-control';
-import {
-  formatRemoteControlAlreadyRunning,
-  inspectRemoteControlLock,
-} from '#/cli/sub/web/remote-control-lock';
 import { formatReadyBanner, startServerForeground } from '#/cli/sub/web/run';
 import { parseServerOptions, tryResolveServerToken } from '#/cli/sub/web/shared';
 import { openUrl } from '#/utils/open-url';
@@ -72,6 +71,7 @@ export async function handleRemoteControlCommand(host: SlashCommandHost): Promis
             homeDir: dataDir,
             localOrigin: origin,
             localServerToken: token,
+            clientVersion: `kimi-code/${getVersion()}`,
             onStatus,
           });
           const url = buildRemoteControlUrl(remoteControl.deviceId, session?.id);
