+++ packages/kap-server/src/protocol/error-codes.ts
@@ -61,6 +61,7 @@ export const ErrorCode = {
   CAPABILITY_UNSUPPORTED: 40925,
   RUNTIME_UNAVAILABLE: 40926,
   PROMPT_ID_CONFLICT: 40927,
+  REMOTE_CONTROL_ALREADY_RUNNING: 40928,
   MCP_OAUTH_FAILED: 40929,
 
   APPROVAL_EXPIRED: 41001,
