+++ packages/kap-server/src/start.ts
@@ -81,13 +81,11 @@ import { TranscriptService } from './services/transcript/transcriptService';
 import { ModelCatalogRefreshScheduler } from './services/modelCatalog/modelCatalogRefreshScheduler';
 import { startConfigChangedPublisher } from './services/config/configChangedPublisher';
 import { createAuthFailureLimiter } from './middleware/rateLimit';
-import {
-  createAuthTokenService,
-  type IAuthTokenService,
-} from './services/auth/authTokenService';
+import { createAuthTokenService, type IAuthTokenService } from './services/auth/authTokenService';
 import { createCredentialValidator } from './services/auth/credentials';
 import { resolvePasswordHash } from './services/auth/password';
 import { createTokenStore } from './services/auth/tokenStore';
+import { createRemoteControlService } from './services/remoteControl/remoteControlService';
 
 import { drainGlobalSearchDisposals, IGlobalSearchService } from './search/searchService';
 
@@ -193,6 +191,14 @@ export async function startServer(opts: ServerStartOptions): Promise<RunningServ
   }
   const validateCredential = createCredentialValidator(authTokenService, opts.rpcToken);
   const logging = resolveLoggingConfig({ homeDir, env: process.env });
+  let boundPort = port;
+  const remoteControlService = createRemoteControlService({
+    homeDir,
+    localOrigin: () => `http://${host}:${boundPort}`,
+    localServerToken: () => authTokenService.getToken(),
+    clientVersion: `kimi-code/${serverVersion}`,
+    logger,
+  });
   const { app: core } = bootstrap(
     {
       homeDir,
@@ -292,6 +298,7 @@ export async function startServer(opts: ServerStartOptions): Promise<RunningServ
 
   const close = async (): Promise<void> => {
     configChangedPublisher.close();
+    await remoteControlService.close();
     await app.close();
     configWarningSubscription.dispose();
     pluginChangeSubscription.dispose();
@@ -404,6 +411,7 @@ export async function startServer(opts: ServerStartOptions): Promise<RunningServ
           { name: 'terminals', description: 'PTY terminal sessions' },
           { name: 'fs', description: 'Filesystem operations' },
           { name: 'files', description: 'File upload & download' },
+          { name: 'remote-control', description: 'Remote Control tunnel' },
         ],
       },
       transformObject: (documentObject) => {
@@ -434,6 +442,15 @@ export async function startServer(opts: ServerStartOptions): Promise<RunningServ
       opts.pluginMarketplaceUrl === undefined &&
       (process.env['KIMI_CODE_PLUGIN_MARKETPLACE_URL'] === undefined ||
         process.env['KIMI_CODE_PLUGIN_MARKETPLACE_FROM_DEV_SERVER'] === '1'),
+    remoteControl: {
+      service: remoteControlService,
+      staticEnableError:
+        exposureClass !== 'loopback'
+          ? 'Remote Control requires a loopback host.'
+          : opts.disableAuth === true
+            ? 'Remote Control cannot be combined with --dangerous-bypass-auth.'
+            : undefined,
+    },
     onShutdown: () => {
       void close().catch((err: unknown) => logger.error({ err }, 'server close failed'));
     },
@@ -568,7 +585,7 @@ export async function startServer(opts: ServerStartOptions): Promise<RunningServ
   }
 
   const address = app.server.address();
-  const boundPort = typeof address === 'object' && address !== null ? address.port : port;
+  boundPort = typeof address === 'object' && address !== null ? address.port : port;
   await registration.update({ port: boundPort });
 
   void modelCatalogRefreshScheduler.start().catch((error) => {
