+++ packages/remote-control/src/index.ts
@@ -1,0 +1,2 @@
+export * from './remote-control';
+export * from './lock';
